interleave multiple text comparisions in lcs simd
Interleaving three text comparisions in the LCS Simd implementation
appears to max out the throughput.
According to my benchmarks this gives the following speedups for Ascii texts:
len  8: +3%
len 16: +35%
len 32: +36%
len 64: +49%

This suggests that for len 8 texts we are already bottlenecked in a different place.
This is not all to surprising, since on my CPU we currently need around 0.75 cycles
per 8 character text comparision.